third_party/blink/web_tests/platform/generic/fast/mediastream/getusermedia-expected.txt

CONSOLE WARNING: Unknown constraint named valid_but_unsupported_1 rejected
Tests webkitGetUserMedia.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.isUseCounted(document, GetUserMediaPrefixed) is false
PASS internals.isUseCounted(document, GetUserMediaLegacy) is false
PASS navigator.webkitGetUserMedia({audio:false, video:false}, error, error); threw exception TypeError: Failed to execute 'webkitGetUserMedia' on 'Navigator': At least one of audio and video must be requested.
PASS navigator.webkitGetUserMedia({audio:true}, gotStreamNoMore, error); did not throw exception.
PASS navigator.getUserMedia({audio:false, video:false}, error, error); threw exception TypeError: Failed to execute 'getUserMedia' on 'Navigator': At least one of audio and video must be requested.
PASS navigator.getUserMedia({audio:true}, gotStream1, error); did not throw exception.
PASS internals.isUseCounted(document, GetUserMediaPrefixed) is true
PASS internals.isUseCounted(document, GetUserMediaLegacy) is true
PASS internals.isUseCounted(document, GetUserMediaPromise) is false
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 0
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 0
PASS navigator.getUserMedia({video:true}, gotStream2, error); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 0
PASS stream.getVideoTracks().length is 1
PASS navigator.getUserMedia({audio:true, video:true}, gotStream3, error); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS navigator.getUserMedia({audio:{mandatory:{}, optional:[]}, video:true}, gotStream4, error); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS Error callback called.
PASS errorArg.name is "OverconstrainedError"
PASS errorArg.constraint is "valid_but_unsupported_1"
PASS navigator.getUserMedia({audio:{mandatory:{'valid_and_supported_1':1}, optional:[{'valid_but_unsupported_1':0}]}, video:true}, gotStream5, error); did not throw exception.
PASS navigator.getUserMedia({audio:{mandatory:{'valid_but_unsupported_1':0}, optional:[]}, video:true}, gotStreamInError, error1); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS successfullyParsed is true

TEST COMPLETE
